create opt etcd directory workaround for issue 11
Issue 11
stx-postrun.bbclass: Once the ansible-playbook runs it resets ETCD_DATA_DIR
to /opt/etcd/19.01/controller.etcd (at least on the centos version) in
/etc/etcd/etcd.conf. This directory does not exist and consequently etcd fails
to start. This workaround creates the data directory with the correct ownership.

The actual fix is the failure in creating the directory and fix it there. Leaving
the issue open.