erlang: fix to build with openssl10
erlang R18B03 is not compatibel with openssl 1.1.x, previously
the ssl reladted features were disabled, but it's required by
rabbitmq-server 3.6.5 which is used by stx 3.0.

To align with stx 3.0, we need to build erlang with compatible
openssl version so that we can upgrade rabbitmq-server to 3.6.5.

The commit includes the following changes to that goal:
- Add dependency on openssl10 1.0.2 instead of openssl 1.1.1d
- Add bbappend for openssl10 1.0.2 to handle sysroot conflicts
  with openssl 1.1.1d.
- Handle the openssl10 lib and herders in do_configure_prepend
  so erlang builds successfaully with ssl enabled.
- Fix two QA warnings.

fix #250

Signed-off-by: Jackie Huang <jackie.huang@windriver.com>